% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DEMIExperiment-methods.R
\name{checkDEMIExperiment_experiment}
\alias{checkDEMIExperiment_experiment}
\title{Checks if the \code{experiment} is correct}
\usage{
checkDEMIExperiment_experiment(experiment)
}
\arguments{
\item{experiment}{A \code{character}.}
}
\value{
Returns NULL if \code{experiment} is ok, else returns an error message.
}
\description{
Checks if the \code{experiment} is correct for the DEMI analysis. It is used internally in DEMI
analysis.
}
\author{
Sten Ilmjarv
}

