\name{checkDEMIExperiment_maxprobes}
\alias{checkDEMIExperiment_maxprobes}
\title{Checks if \code{maxprobes} is correct}
\usage{
  checkDEMIExperiment_maxprobes(maxprobes)
}
\arguments{
  \item{maxprobes}{A \code{numeric}.}
}
\value{
  Returns NULL if \code{maxprobes} is ok, else returns an
  error message.
}
\description{
  Checks if \code{maxprobes} is correct for the DEMI
  analysis. It is used internally in DEMI analysis.
}
\author{
  Sten Ilmjarv
}

