\name{checkDEMIExperiment_maxtargets}
\alias{checkDEMIExperiment_maxtargets}
\title{Checks if \code{maxtargets} is correct}
\usage{
  checkDEMIExperiment_maxtargets(maxtargets)
}
\arguments{
  \item{maxtargets}{A \code{numeric}.}
}
\value{
  Returns NULL if \code{maxtargets} is ok, else returns an
  error message.
}
\description{
  Checks if \code{maxtargets} is correct for the DEMI
  analysis. It is used internally in DEMI analysis.
}
\author{
  Sten Ilmjarv
}

