\docType{methods}
\name{cleanorganismname}
\alias{cleanorganismname}
\title{Cleans the organism name from redundant characters}
\usage{
  cleanorganismname(organism)
}
\arguments{
  \item{organism}{A \code{character}. A \code{character}
  specifing the organism name.}
}
\value{
  A \code{character} representing clean organism name.
}
\description{
  Cleans the organism name from redundant. It is used
  internally in DEMI analysis.
}
\author{
  Sten Ilmjarv
}

