\name{DEMICel-class}
\alias{DEMICel-class}
\title{Class \code{DEMICel}}
\arguments{
  \item{celMatrix}{A \code{matrix}. The raw expression
  matrix.}

  \item{normMatrix}{A \code{matrix}. The normalized
  expression matrix.}
}
\description{
  The class \code{DEMICel} holds the raw and normalized
  expression matrices.
}
\author{
  Sten Ilmjarv
}

