% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DEMIClust-class}
\alias{DEMIClust-class}
\title{Class \code{DEMIClust}}
\arguments{
\item{experiment}{A \code{DEMIExperiment} object. Holds the \code{DEMIExperiment} object whose metadata
(such as normalized expression values) is used to cluster the probes.}

\item{group}{A \code{DEMIGroup} object. Defines the groups that are used for clustering. The \code{DEMIGroup}
object uses CEL file names to determine which files belong to which group. It uses
\code{grep} function to locate the group names from the CEL file names.}

\item{clust.method}{A \code{function}. Defines the function used for clustering. The user can
build a custom clustering function. The input of the custom function needs
to be the same \code{DEMIClust} object and the output is a list of probes, where
each list corresponds to a specific cluster. The default function is \code{demi.wilcox.test} that
utilizes \code{wilcox.test}.}

\item{cluster}{A \code{list}. Holds the probes of different clusters in a \code{list}.}

\item{cutoff.pvalue}{A \code{numeric}. Sets the cut-off p-value used for determining statistical
          significance of the probes when clustering the probes into clusters. Default is 0.05.}
}
\description{
The class \code{DEMIClust} stores the probe clusters in a \code{DEMIClust} object.
}
\author{
Sten Ilmjarv
}

