\name{DEMIGroup-class}
\alias{DEMIGroup-class}
\title{Class \code{DEMIGroup}}
\arguments{
  \item{groupA}{A \code{character}. Holds the name of group
  A.}

  \item{groupB}{A \code{character}. Holds the name of group
  B.}

  \item{indexA}{A \code{numeric}. A vector of column
  indexes belonging to group A.}

  \item{indexB}{A \code{numeric}. A vector of column
  indexes belonging to group B.}

  \item{groupNames}{A \code{character}. Holds the names of
  custom groups created by the user.}
}
\description{
  The class \code{DEMIGroup} holds the information about
  the groups. Such as the column indexes of both groups and
  names of the groups.
}
\author{
  Sten Ilmjarv
}

