\docType{data}
\name{DEMIMessages}
\alias{DEMIMessages}
\title{A \code{list} of DEMI messages}
\format{List of 33
 $ demiequal                  :List of 3
 $ demi.wilcox.test.fast      :List of 3
 $ demi.wilcox.test           :List of 3
 $ demiExperiment_t.test      :List of 1
 $ demi.comp.test             :List of 1
 $ takestime                  :function (no.of.probes)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 15 41 131 29 145 41 41
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ probesdone                 :function (no.of.probes)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 42 16 42 99 30 113 42 42
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ sectionsizeMissing         :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 49 24 49 118 38 132 49 49
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ positiveIntegerOrZero      :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 50 27 50 117 41 131 50 50
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ sectionsizeUse             :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 52 20 52 107 34 121 52 52
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ parameterOfClassMissing    :function (parameter, cls)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 53 29 53 130 43 144 53 53
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ paramAndParamNotEqualLength:function (param1, param2)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 54 33 54 142 47 156 54 54
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ wrongDefinition            :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 55 21 58 94 35 150 55 58
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ savingWhatTo               :function (what, to)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 59 18 59 82 32 96 59 59
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ customTargets              :List of 11
 $ DEMIPathway                :List of 2
 $ DEMIClust                  :List of 4
 $ DEMIDiff                   :List of 12
 $ DEMIExperiment             :List of 31
 $ DEMIResults                :List of 2
 $ diffexp                    :List of 2
 $ normalization              :List of 3
 $ parameterMissing           :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 171 22 171 93 36 107 171 171
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ parameterNotOfClass        :function (parameter, cls)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 172 25 172 116 39 130 172 172
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ isNotError                 :function (parameter, what)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 173 16 173 93 30 107 173 173
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ hasToBeNumericBetween      :function (parameter, start, end)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 174 27 174 142 41 156 174 174
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ tooManyParameters          :function (parameter, len)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 175 23 175 111 37 125 175 175
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ positiveInteger            :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 176 21 176 106 35 120 176 176
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ positiveIntegerUpTo        :function (parameter, to)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 177 25 177 124 39 138 177 177
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ hasToBeFunction            :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 178 21 178 98 35 112 178 178
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ notEmptyString             :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 179 20 179 101 34 115 179 179
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ done                       :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 180 10 180 32 24 46 180 180
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418> 
 $ demiStart                  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 184 15 195 111 29 139 184 195
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x9c1b418>}
\usage{
  DEMIMessages
}
\description{
  This \code{list} contains all the messages shown to the
  user when running DEMI analysis. It is used internally in
  DEMI analysis.
}
\author{
  Sten Ilmjarv, Hendrik Luuk
}
\keyword{data}
\keyword{datasets}

