\docType{methods}
\name{diffexp}
\alias{diffexp}
\alias{diffexp,DEMIDiff-method}
\title{Initializes the differential expression analysis}
\arguments{
  \item{object}{A \code{DEMIDiff} object.}
}
\value{
  Returns the \code{DEMIDiff} object that is updated with
  the results from differential expression analysis.
}
\description{
  The function \code{diffexp} performs differential
  expression analysis. This function is used internally by
  DEMI analysis.
}
\author{
  Sten Ilmjarv
}
\seealso{
  \code{DEMIDiff}
}

