\docType{methods}
\name{diffSpliceScore}
\alias{diffSpliceScore}
\title{Calculate differential splice scores}
\usage{
  diffSpliceScore(x)
}
\arguments{
  \item{x}{A \code{data.frame}.}
}
\value{
  Returns the differential splice score as \code{numeric}.
}
\description{
  The function \code{diffSpliceScore} calculates
  differential splice scores in DEMI analysis. It is used
  internally in DEMI analysis. In the current
  implementation of DEMI it is not used.
}
\author{
  Sten Ilmjarv
}

