% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DEMIClust-methods.R, R/DEMIDiff-methods.R, R/DEMIExperiment-methods.R
\docType{methods}
\name{getExperiment}
\alias{getExperiment}
\alias{getExperiment,DEMIClust-method}
\alias{getExperiment,DEMIDiff-method}
\alias{getExperiment,DEMIExperiment-method}
\title{Returns the \code{experiment} parameter}
\usage{
getExperiment(object)

\S4method{getExperiment}{DEMIClust}(object)

\S4method{getExperiment}{DEMIDiff}(object)

\S4method{getExperiment}{DEMIExperiment}(object)
}
\arguments{
\item{object}{A \code{DEMIExperiment}, \code{DEMIClust} or \code{DEMIDiff} object.}
}
\value{
Returns the \code{experiment} parameter. If the input object is \code{DEMIExperiment} it returns a
		   character, if the input object is either \code{DEMIClust} or \code{DEMIDiff} it returns a
		   \code{DEMIExperiment} object.
}
\description{
Returns the \code{experiment} parameter of the specified object. For object of class \code{DEMIExperiment}
it returns the name given to the experiment. For objects of class \code{DEMIClust} or \code{DEMIDiff}
it returns the initial \code{DEMIExperiment} object. This function can be useful if the user wants
to access metadata, such as annotations and alignments from other DEMI analysis objects. As well
as accessing the name of the analysis.
}
\examples{
\dontrun{

# To use the example we need to download a subset of CEL files from
# http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE9819 published
# by Pradervand et al. 2008.

# Set the destination folder where the downloaded files fill be located.
# It can be any folder of your choosing.
destfolder <- "demitest/testdata/"

# Download packed CEL files and change the names according to the feature
# they represent (for example to include UHR or BRAIN in them to denote the
# features).
# It is good practice to name the files according to their features which
# allows easier identification of the files later.

ftpaddress <- "ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM247nnn"
download.file( paste( ftpaddress, "GSM247694/suppl/GSM247694.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "UHR01_GSM247694.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247695/suppl/GSM247695.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "UHR02_GSM247695.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247698/suppl/GSM247698.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "UHR03_GSM247698.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247699/suppl/GSM247699.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "UHR04_GSM247699.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247696/suppl/GSM247696.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "BRAIN01_GSM247696.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247697/suppl/GSM247697.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "BRAIN02_GSM247697.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247700/suppl/GSM247700.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "BRAIN03_GSM247700.CEL.gz", sep = "" ) )
download.file( paste( ftpaddress, "GSM247701/suppl/GSM247701.CEL.gz", sep = "/" ),
		destfile = paste( destfolder, "BRAIN04_GSM247701.CEL.gz", sep = "" ) )

# We need the gunzip function (located in the R.utils package) to unpack the gz files.
# Also we will remove the original unpacked files for we won't need them.
library( R.utils )
for( i in list.files( destfolder ) ) {
	gunzip( paste( destfolder, i, sep = "" ), remove = TRUE )
}

# Now we can continue the example of the function getExperiment

# Set up an experiment
demiexp <- DEMIExperiment( analysis = 'gene', celpath = destfolder,
			experiment = 'myexperiment', organism = 'homo_sapiens' )

# Create clusters with an optimized wilcoxon's rank sum test incorporated within demi that
# precalculates the probabilities
demiclust <- DEMIClust( demiexp, group = c( "BRAIN", "UHR" ), clust.method = demi.wilcox.test.fast )

# Calcuate differential expression
demidiff <- DEMIDiff( demiclust )

# Retrieve the 'experiment' parameter
getExperiment( demiexp )
getExperiment( demiclust )
getExperiment( demidiff )

}
}
\author{
Sten Ilmjarv
}
\seealso{
\code{DEMIExperiment}, \code{DEMIClust}, \code{DEMIDiff}
}

