\name{initialize.DEMIDiff}
\alias{initialize.DEMIDiff}
\title{Initializes the \code{DEMIDiff} object}
\usage{
  initialize.DEMIDiff(.Object, ...)
}
\arguments{
  \item{.Object}{A DEMIDiff object.}

  \item{...}{Additional arguments that may never be used.}
}
\value{
  Returns a 'DEMDiff' object.
}
\description{
  Initializes the \code{DEMIDiff} object.
}
\author{
  Sten Ilmjarv
}

