\name{initialize.DEMIExperiment}
\alias{initialize.DEMIExperiment}
\title{Initializes the \code{DEMIExperiment} object}
\usage{
  initialize.DEMIExperiment(.Object, ...)
}
\arguments{
  \item{.Object}{A DEMIExperiment object.}

  \item{...}{Additional arguments that may never be used.}
}
\value{
  Returns a 'DEMIExperiment' object.
}
\description{
  Initializes the \code{DEMIExperiment} object.
}
\author{
  Sten Ilmjarv
}

