\name{initialize.DEMIGroup}
\alias{initialize.DEMIGroup}
\title{Initializes the \code{DEMIGroup} object}
\usage{
  initialize.DEMIGroup(.Object, ...)
}
\arguments{
  \item{.Object}{A DEMIGroup object.}

  \item{...}{Additional arguments that may never be used.}
}
\value{
  Returns a 'DEMIGroup' object.
}
\description{
  Initializes the \code{DEMIGroup} object.
}
\author{
  Sten Ilmjarv
}

