\name{initialize.DEMIResult}
\alias{initialize.DEMIResult}
\title{Initializes the \code{DEMIResult} object}
\usage{
  initialize.DEMIResult(.Object, ...)
}
\arguments{
  \item{.Object}{A DEMIResult object.}

  \item{...}{Additional arguments that may never be used.}
}
\value{
  Returns a 'DEMIResult' object.
}
\description{
  Initializes the \code{DEMIResult} object.
}
\author{
  Sten Ilmjarv
}

