% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DEMIExperiment-methods.R
\docType{methods}
\name{loadCytoband}
\alias{loadCytoband}
\alias{loadCytoband,DEMIExperiment,environment-method}
\title{Loads the karyotype information specified by the \code{DEMIExperiment} object}
\usage{
loadCytoband(object, pkg)

\S4method{loadCytoband}{DEMIExperiment,environment}(object, pkg)
}
\arguments{
\item{object}{A \code{DEMIExperiment} object.}

\item{pkg}{An \code{environment}. Specifies the environment where to load the data from.}
}
\value{
Returns a \code{data.frame} with karyotype information.
}
\description{
The function \code{loadCytoband} loads the karyotype information for the specified \code{DEMIExperiment} object. It
is used internally in DEMI analysis.
}
\author{
Sten Ilmjarv
}
\seealso{
\code{DEMIExperiment}, \code{environment}
}

