\name{validDEMIClust}
\alias{validDEMIClust}
\title{Validates the \code{DEMIClust} object}
\usage{
  validDEMIClust(object)
}
\arguments{
  \item{object}{A \code{DEMIClust} object.}
}
\value{
  Returns a validated \code{DEMIClust} object.
}
\description{
  Validates the \code{DEMIClust} object.
}
\author{
  Sten Ilmjarv
}

