\name{validDEMIExperiment}
\alias{validDEMIExperiment}
\title{Validates the \code{DEMIExperiment} object}
\usage{
  validDEMIExperiment(object)
}
\arguments{
  \item{object}{A \code{DEMIExperiment} object.}
}
\value{
  Returns a validated \code{DEMIExperiment} object.
}
\description{
  Validates the \code{DEMIExperiment} object
}

