% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/normalization-methods.R
\docType{methods}
\name{celMatrixNormalize}
\alias{celMatrixNormalize}
\alias{celMatrixNormalize,DEMIExperiment,function-method}
\title{Initializes the normalization of the raw expression matrix}
\usage{
celMatrixNormalize(object, fun)

\S4method{celMatrixNormalize}{DEMIExperiment,`function`}(object, fun)
}
\arguments{
\item{object}{A \code{DEMIExperiment} object. It stores the raw expression matrix.}

\item{fun}{A \code{function}. The function used to normalize the raw expression matrix.}
}
\value{
Returns a \code{DEMIExperiment} object updated with normalized expression matrix.
}
\description{
The function \code{celMatrixNormalize} initializes the normalization of the raw expression matrix
in the \code{DEMIExperiment} object. It is used internally in DEMI analysis.
}
\author{
Sten Ilmjarv
}

