% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/other-methods.R
\docType{methods}
\name{cleanorganismname}
\alias{cleanorganismname}
\title{Cleans the organism name from redundant characters}
\usage{
cleanorganismname(organism)
}
\arguments{
\item{organism}{A \code{character}. A \code{character} specifing the organism name.}
}
\value{
A \code{character} representing clean organism name.
}
\description{
Cleans the organism name from redundant. It is used internally in DEMI
analysis.
}
\author{
Sten Ilmjarv
}

