\docType{data}
\name{DEMIMessages}
\alias{DEMIMessages}
\title{A \code{list} of DEMI messages}
\format{List of 33
 $ demiequal                  :List of 3
 $ demi.wilcox.test.fast      :List of 3
 $ demi.wilcox.test           :List of 3
 $ demiExperiment_t.test      :List of 1
 $ demi.comp.test             :List of 1
 $ takestime                  :function (no.of.probes)  
 $ probesdone                 :function (no.of.probes)  
 $ sectionsizeMissing         :function ()  
 $ positiveIntegerOrZero      :function (parameter)  
 $ sectionsizeUse             :function ()  
 $ parameterOfClassMissing    :function (parameter, cls)  
 $ paramAndParamNotEqualLength:function (param1, param2)  
 $ wrongDefinition            :function ()  
 $ savingWhatTo               :function (what, to)  
 $ customTargets              :List of 11
 $ DEMIPathway                :List of 2
 $ DEMIClust                  :List of 4
 $ DEMIDiff                   :List of 12
 $ DEMIExperiment             :List of 31
 $ DEMIResults                :List of 2
 $ diffexp                    :List of 2
 $ normalization              :List of 3
 $ parameterMissing           :function (parameter)  
 $ parameterNotOfClass        :function (parameter, cls)  
 $ isNotError                 :function (parameter, what)  
 $ hasToBeNumericBetween      :function (parameter, start, end)  
 $ tooManyParameters          :function (parameter, len)  
 $ positiveInteger            :function (parameter)  
 $ positiveIntegerUpTo        :function (parameter, to)  
 $ hasToBeFunction            :function (parameter)  
 $ notEmptyString             :function (parameter)  
 $ done                       :function ()  
 $ demiStart                  :function ()}
\usage{
  DEMIMessages
}
\description{
  This \code{list} contains all the messages shown to the
  user when running DEMI analysis. It is used internally in
  DEMI analysis.
}
\author{
  Sten Ilmjarv, Hendrik Luuk
}
\keyword{data}
\keyword{datasets}

