% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{DEMIMessages}
\alias{DEMIMessages}
\title{A \code{list} of DEMI messages}
\format{\preformatted{List of 33
 $ demiequal                  :List of 3
  ..$ main             : chr "# Clustering probes into 'equal' cluster based on Wilcoxon rank sum test if the groups are bigger than 2\\n"
  ..$ custom.approach  : chr "\\tUsing an optimized implementation to perform Wilcoxon rank sum test\\n"
  ..$ standard.approach: chr "\\tUsing the native implementation of the Wilcoxon rank sum test (function 'wilcox.test')\\n"
 $ demi.wilcox.test.fast      :List of 3
  ..$ main             : chr "# Clustering probes into 'higher' and 'lower' clusters based on Wilcoxon rank sum test\\n"
  ..$ custom.approach  : chr "\\tUsing an optimized implementation to perform Wilcoxon rank sum test\\n"
  ..$ standard.approach: chr "\\tUsing the native implementation of the Wilcoxon rank sum test (function 'wilcox.test')\\n"
 $ demi.wilcox.test           :List of 3
  ..$ main       : chr "# Clustering probes into 'higher' and 'lower' clusters based on Wilcoxon rank sum test\\n"
  ..$ exact.false: chr "\\tUsing wilcox with parameter 'exact = FALSE' since sample sizes are sufficiently large.\\n"
  ..$ exact.true : chr "\\tUsing wilcox with parameter 'exact = TRUE'.\\n"
 $ demiExperiment_t.test      :List of 1
  ..$ main: chr "# Clustering probes into 'higher' and 'lower' clusters based on t-test (function 't.test')\\n"
 $ demi.comp.test             :List of 1
  ..$ main: chr "# Clustering probes into 'higher' and 'lower' clusters\\n"
 $ takestime                  :function (no.of.probes)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 15 41 131 29 145 41 41
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ probesdone                 :function (no.of.probes)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 42 16 42 99 30 113 42 42
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ sectionsizeMissing         :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 49 24 49 118 38 132 49 49
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ positiveIntegerOrZero      :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 50 27 50 117 41 131 50 50
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ sectionsizeUse             :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 52 20 52 107 34 121 52 52
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ parameterOfClassMissing    :function (parameter, cls)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 53 29 53 130 43 144 53 53
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ paramAndParamNotEqualLength:function (param1, param2)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 54 33 54 142 47 156 54 54
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ wrongDefinition            :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 55 21 58 94 35 150 55 58
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ savingWhatTo               :function (what, to)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 59 18 59 82 32 96 59 59
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ customTargets              :List of 11
  ..$ whatMissingFrom    :function (what, from)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 64 23 64 132 51 160 64 64
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ pmsizeMissing      : chr "# Note: column 'pmsize' is unspecified in the 'blat' object. We will therefore assume that all probe matches to the target are "| __truncated__
  ..$ startMissing       : chr "# Note: column 'start' is unspecified in the 'blat' object.\\n"
  ..$ strandMissing      : chr "# Note: column 'strand' is unspecified in the 'blat' object. We will therefore assume that all probe matches were on the major "| __truncated__
  ..$ probesNotPresent   :function (notpresent)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 68 24 68 198 52 226 68 68
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ geneMissing        : chr "# Note: column 'geneID' is missing from the 'anno' object. Will duplicate 'targetID' as 'geneID' \\n"
  ..$ annoMissing        : chr "# Note: 'anno' has not been specified. Creating custom annotation\\n"
  ..$ ignoreStrand       :function (strand)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 71 20 71 111 48 139 71 71
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ pmsizeSmallerRemove: chr "# Note: Removing probes with perfect match size less than indicated in the DEMIExperiment object\\n"
  ..$ determineMatches   : chr "# Determining the number of hits on each target for every probe\\n"
  ..$ error              : chr "You can't add custom region alignments's and annotation info to the exon analysis. You can only add them to 'transcript' and 'g"| __truncated__
 $ DEMIPathway                :List of 2
  ..$ main   : chr "# Performing pathway analysis of differentially expressed genes\\n"
  ..$ cantrun: chr "DEMIPathway can only be run if the analysis type in DEMIExperimebt object is a \\"gene\\" or \\"transcript\\"\\n"
 $ DEMIClust                  :List of 4
  ..$ noCELFilesWithGroupname:function (groupName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 84 31 85 136 59 192 84 85
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ usingBuiltIn           : chr "# Note: Clustering with demi built-in functions\\n"
  ..$ usingCreaterOrLower    : chr "# Note: A heuristic-based method will be used to assess probe-level differential expression as one of the group sizes is <= 2.\\"| __truncated__
  ..$ usingUserProvided      : chr "# Note: Clustering with user-provided function\\n"
 $ DEMIDiff                   :List of 12
  ..$ zeroTargetsFound      : chr "0 specified targets were found in the experiment"
  ..$ calcDiffExp           : chr "# Calculating differential expression "
  ..$ betweenGroups         :function (groupA, groupB)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 96 21 96 178 49 206 96 96
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ onUserDefined         : chr "on user-defined clusters\\n"
  ..$ analyzingClusterNative:function (clusterName, groupA, groupB)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 98 30 98 246 58 274 98 98
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ analyzingClusterCustom:function (clusterName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 99 30 99 135 58 163 99 99
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ calcHyperGeo          : chr "\\t\\tCalculating hypergeometric distribution p-values ... "
  ..$ annotatingResults     : chr "\\t\\tAnnotating results ..."
  ..$ multipleCorrectionBH  : chr "\\t\\tAdjusting p-values for multiple testing by the FDR method (Benjamini & Hochberg, 1995) ... "
  ..$ multipleCorrectionBY  : chr "\\t\\tAdjusting p-values for multiple testing by the FDR method under dependence (Benjamini & Yekutieli, 2001) ... "
  ..$ multipleCorrectionBONF: chr "\\t\\tAdjusting p-values for multiple testing by the Bonferroni method ... "
  ..$ diffExpDone           : chr "# Calculating differential expression ... DONE\\n"
 $ DEMIExperiment             :List of 31
  ..$ invalidWhatChooseFrom   :function (what, chooseFrom)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 112 29 112 188 57 216 112 112
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ notFolder               :function (parameter)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 113 17 113 91 45 119 113 113
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ folderEmpty             :function (parameter)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 114 19 114 130 47 158 114 114
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ notACELFile             :function (file, parameter)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 115 19 115 135 47 163 115 115
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ maxprobesError          : chr "The parameter 'maxprobes' has to be a positive integer or set as 'median', 'max' or not defined at all which by default set's i"| __truncated__
  ..$ probesExcluded          :function (excludedProbes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 117 22 117 195 50 223 117 117
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ noTargetsFound          : chr "No targets were found"
  ..$ noProbesMeasuring       : chr "No probes measuring the specified targets matched the experiment criteria"
  ..$ attachErrorResultsExists:function (variable)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 120 32 120 288 60 316 120 120
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ attachErrorDiffDEMIExp  : chr "Can't add results of class 'DEMIDiff' to this experiment of class 'DEMIExperiment' because they represent different 'DEMIExperi"| __truncated__
  ..$ loadingCEL              : chr "# Loading CEL files and creating expression matrix\\n"
  ..$ includedCELFiles        :function (celfiles)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 123 24 123 115 52 143 123 123
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ usingMicroarrayPlatform :function (platform)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 124 31 124 120 59 148 124 124
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ libraryNotInstalled     :function (packageName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 125 27 125 114 55 142 125 125
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ searchingInternetRepo   :function (packageName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 126 29 126 142 57 170 126 126
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ packageNotFoundFromRepo :function (packageName, package_url)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 127 31 127 143 59 171 127 127
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ installingPackage       :function (packageName, repository)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 128 25 128 142 53 170 128 128
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ checkYourInternet       : chr "The current operation could not access the internet. Please check your internet connectction"
  ..$ loadingAnnoSuccess      :function (packageName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 130 26 130 133 54 161 130 130
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ loadingAnnoFail         :function (annoTableName, packageName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 131 23 131 179 51 207 131 131
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ sectionsizeNotAvail     :function (sectionsizes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 132 27 132 169 55 197 132 132
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ cantLoadWhatTable       :function (whatTable)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 133 25 133 116 53 144 133 133
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ loadingBlatSuccess      :function (packageName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 134 26 134 132 54 160 134 134
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ blatDoesNotExists       :function (blat, platform, packageName)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 135 25 135 192 53 220 135 135
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ pmsizeNotAvail          :function (pmsizes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 136 22 136 182 50 210 136 136
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ ignoreStrandMatches     :function (strand)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 137 27 137 112 55 140 137 137
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ loadingCytoband         : chr "# Loading cytoband information\\n"
  ..$ loadingPathway          : chr "# Loading pathway information\\n"
  ..$ check4probeError        :function (difference)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 140 24 142 124 52 173 140 142
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ check4targetError       :function (notfound)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 143 25 143 181 53 209 143 143
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
  ..$ zeroTargetsFound        : chr "0 specified targets were found in the experiment"
 $ DEMIResults                :List of 2
  ..$ resultsContain  : chr "the results list can only contain objects of class 'DEMIResult'"
  ..$ resultsUndefined: chr "no results have been defined"
 $ diffexp                    :List of 2
  ..$ matchesCluster: chr "\\t\\tCalculating matches for probes in cluster ... "
  ..$ matchesAll    : chr "\\t\\tCalculating matches over all probes ... "
 $ normalization              :List of 3
  ..$ main        : chr "# Normalizing expression values "
  ..$ normrrank   : chr " - using 'relative rank' as the normalization method\\n"
  ..$ normquantile: chr "- using 'quantile normalization' as the normalization method\\n"
 $ parameterMissing           :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 171 22 171 93 36 107 171 171
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ parameterNotOfClass        :function (parameter, cls)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 172 25 172 116 39 130 172 172
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ isNotError                 :function (parameter, what)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 173 16 173 93 30 107 173 173
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ hasToBeNumericBetween      :function (parameter, start, end)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 174 27 174 142 41 156 174 174
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ tooManyParameters          :function (parameter, len)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 175 23 175 111 37 125 175 175
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ positiveInteger            :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 176 21 176 106 35 120 176 176
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ positiveIntegerUpTo        :function (parameter, to)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 177 25 177 124 39 138 177 177
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ hasToBeFunction            :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 178 21 178 98 35 112 178 178
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ notEmptyString             :function (parameter)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 179 20 179 101 34 115 179 179
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ done                       :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 180 10 180 32 24 46 180 180
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
 $ demiStart                  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 184 15 195 111 29 139 184 195
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f8e9424f740> 
}}
\usage{
DEMIMessages
}
\description{
This \code{list} contains all the messages shown to the user when running DEMI analysis. It
is used internally in DEMI analysis.
}
\author{
Sten Ilmjarv, Hendrik Luuk
}
\keyword{data}

