% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DEMIDiff-methods.R
\docType{methods}
\name{diffexp}
\alias{diffexp}
\alias{diffexp,DEMIDiff-method}
\title{Initializes the differential expression analysis}
\usage{
diffexp(object)

\S4method{diffexp}{DEMIDiff}(object)
}
\arguments{
\item{object}{A \code{DEMIDiff} object.}
}
\value{
Returns the \code{DEMIDiff} object that is updated with the results from differential
		   expression analysis.
}
\description{
The function \code{diffexp} performs differential expression analysis. This function
is used internally by DEMI analysis.
}
\author{
Sten Ilmjarv
}
\seealso{
\code{DEMIDiff}
}

