% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DEMICel-methods.R
\name{initialize.DEMICel}
\alias{initialize.DEMICel}
\title{Initializes a \code{DEMICel} object}
\usage{
initialize.DEMICel(.Object, ...)
}
\arguments{
\item{.Object}{A DEMICel object.}

\item{...}{Additional arguments that may never be used.}
}
\value{
Returns a \code{DEMICel} object
}
\description{
Initializes a \code{DEMICel} object.
}
\author{
Sten Ilmjarv
}

