% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DEMIExperiment-methods.R
\name{initialize.DEMIExperiment}
\alias{initialize.DEMIExperiment}
\title{Initializes the \code{DEMIExperiment} object}
\usage{
initialize.DEMIExperiment(.Object, ...)
}
\arguments{
\item{.Object}{A DEMIExperiment object.}

\item{...}{Additional arguments that may never be used.}
}
\value{
Returns a 'DEMIExperiment' object.
}
\description{
Initializes the \code{DEMIExperiment} object.
}
\author{
Sten Ilmjarv
}

